pyattacker/requirements.txt
base58==2.1.1
bitcoinlib==0.6.14
certifi==2023.11.17
charset-normalizer==3.3.2
ecdsa==0.18.0
fastecdsa==2.3.0
greenlet==3.0.3
idna==3.6
numpy==1.26.3
pandas==2.1.4
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
ripemd-hash==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
